pyproject.toml
[project]
name = "octopilot-mcp"
version = "0.1.0"
description = "MCP server for Octopilot CI/CD toolchain — enables AI agents to detect, generate, and run CI pipelines"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
docker = ["docker>=7.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "C4",  # flake8-comprehensions
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in default args
    "N806",   # variable in function should be lowercase (conflicts with some patterns)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S",    # bandit — security warnings not meaningful in test code
    "N802", # function name should be lowercase (pytest fixtures use uppercase ok)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[project.scripts]
octopilot-mcp = "octopilot_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/octopilot_mcp"]
